+++ spec/swagger_v2/endpoint_versioned_path_spec.rb
@@ -1,17 +1,21 @@
 require 'spec_helper'
 
 describe 'Grape::Endpoint#path_and_definitions' do
-  let(:api) do
-    item = Class.new(Grape::API) do
+  let(:item) do
+    Class.new(Grape::API) do
       version 'v1', using: :path
 
       resource :item do
         get '/'
       end
     end
+  end
+
+  let(:api) do
+    item_api = item
 
     Class.new(Grape::API) do
-      mount item
+      mount item_api
       add_swagger_documentation add_version: true
     end
   end
@@ -27,5 +31,22 @@ describe 'Grape::Endpoint#path_and_definitions' do
 
   it 'tags the endpoint with the resource name' do
     expect(subject.first['/v1/item'][:get][:tags]).to eq ['item']
+  end
+
+  context 'when custom tags are specified' do
+    let(:item) do
+      Class.new(Grape::API) do
+        version 'v1', using: :path
+
+        resource :item do
+          desc 'Item description', tags: ['special-item']
+          get '/'
+        end
+      end
+    end
+
+    it 'tags the endpoint with the custom tags' do
+      expect(subject.first['/v1/item'][:get][:tags]).to eq ['special-item']
+    end
   end
 end